requirements.txt
# Core orchestration and configuration
prefect>=2.20.0
python-dotenv>=1.0.1
pydantic>=2.8.0
websockets>=12.0

# Data processing
pandas>=2.2.2
pyarrow>=16.1.0
requests>=2.32.0

# Google Cloud clients
google-cloud-storage>=2.17.0
google-cloud-bigquery>=3.25.0
google-cloud-pubsub>=2.23.0
google-cloud-secret-manager>=2.20.0

# dbt
dbt-core>=1.8.0
dbt-bigquery>=1.8.0

# Testing and linting
pytest>=8.3.0
pytest-cov>=5.0.0
ruff>=0.5.0
